backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.47",
    "fastapi>=0.128.7",
    "mangum>=0.21.0",
    "openai>=2.20.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
